Clang-Tidy II (#3190)
This is the second batch of changes fixing clang-tidy warnings.

Add CMake option AMReX_CLANG_TIDY_WERROR to update warnings to errors.

Use clang-tidy in CI: GNU@8.4 C++17 Release [lib].

+++ Src/Base/AMReX_FabArrayBase.cpp
@@ -2466,7 +2466,7 @@ FabArrayBase::updateBDKey ()
 bool
 CheckRcvStats (Vector<MPI_Status>& recv_stats, const Vector<std::size_t>& recv_size, int tag)
 {
-    for (int i = 0, n = recv_size.size(); i < n; ++i) {
+    for (int i = 0, n = static_cast<int>(recv_size.size()); i < n; ++i) {
         if (recv_size[i] > 0) {
             std::size_t count = 0;
             int tmp_count = 0;
@@ -2596,7 +2596,7 @@ FabArrayBase::is_cell_centered () const noexcept
 }
 
 bool
-FabArrayBase::isFusingCandidate () const noexcept
+FabArrayBase::isFusingCandidate () const noexcept // NOLINT(readability-convert-member-functions-to-static)
 {
 #ifdef AMREX_USE_GPU
     // This is fine tuned on MI100.
